Expose traversal method independently from middleware
In some cases we may wish to get a list of templates in a directory independently of the the express app/request/response cycle. For example, the build script I'm currently writing to compile static error page html for PEX.

TODO/TOCONSIDER - find a module which does the tree traversal.
