+++ index.js
@@ -1,34 +1,13 @@
 'use strict';
 
 var _ = require('underscore'),
-    fs = require('fs'),
-    path = require('path');
+    traverse = require('./lib/traverse');
 
-module.exports = function (app) {
+var Parser = function (app) {
 
     var ext = '.' + app.get('view engine');
-    var regexp = new RegExp(ext + '$');
 
-    function parseDir(partials, dir, prefix) {
-        var files = fs.readdirSync(dir);
-        prefix = prefix || '';
-
-        _.each(files, function (file) {
-            var stat = fs.statSync(path.join(dir, file)),
-                pathName,
-                partialName;
-            if (stat.isDirectory()) {
-                parseDir(partials, path.join(dir, file), prefix + '/' + file);
-            } else if (file.match(regexp)) {
-                pathName = (prefix + '/' + file).replace(/^\//, '');
-                partialName = pathName.replace(/\//g, '-').replace(regexp, '');
-                partials[partialName] = pathName;
-            }
-        });
-        return partials;
-    }
-
-    var partials = parseDir({}, app.get('views'));
+    var partials = traverse(app.get('views'));
 
     return function (req, res, next) {
         res.locals.partials = res.locals.partials || {};
@@ -36,3 +15,7 @@ module.exports = function (app) {
         next();
     };
 };
+
+Parser.traverse = traverse;
+
+module.exports = Parser;
